+++ lib/app/app_widget.dart
@@ -1,16 +1,15 @@
-import 'package:events4me/app/modules/splash/splash_page.dart';
 import 'package:events4me/app/shared/theme/app_theme.dart';
 import 'package:flutter/material.dart';
-import 'package:get/get_navigation/src/root/get_material_app.dart';
+import 'package:flutter_modular/flutter_modular.dart';
 
 class AppWidget extends StatelessWidget {
   const AppWidget({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return GetMaterialApp(
+    return MaterialApp.router(
       theme: AppTheme().themeData,
-      home: const SplashPage(),
+      routerConfig: Modular.routerConfig,
     );
   }
 }