+++ lib/app/shared/services/dio/error/api_error.dart
@@ -1,0 +1,12 @@
+import 'package:dio/dio.dart';
+
+class ApiError extends DioException {
+  final String apiErrorMessage;
+
+  ApiError(DioException dioError, this.apiErrorMessage)
+      : super(
+            requestOptions: dioError.requestOptions,
+            error: dioError.error,
+            response: dioError.response,
+            type: dioError.type);
+}